Vue Router history mode.
Vue router has 2 mode: hash mode and history mode. The default is hash mode, which make the url like: localhost:8000/#/. so the history mode will make it like: localhost:8000 only.

When make this, the web.php also need to be chaned to get /{any?} and where avoid anything start with graphql, but accept all others.
check regex101 for more information.